add backend processing bits

+++ RetroTrack.Domain/Models/RetroAchievements.cs
@@ -131,7 +131,7 @@ namespace RetroTrack.Domain.Models
         public string Description { get; set; }
 
         [JsonProperty("Points")]
-        public long Points { get; set; }
+        public int Points { get; set; }
 
         [JsonProperty("BadgeName")]
         public string BadgeName { get; set; }
